+++ src/components/Dropdown/Dropdown.stories.tsx
@@ -5,7 +5,7 @@ import { address } from 'faker';
 import Icon from '@mdi/react';
 import { mdiLeaf } from '@mdi/js';
 
-import Dropdown, { OptionProps } from './Dropdown';
+import Dropdown, { DropdownProps, OptionProps } from './Dropdown';
 import variants from '../../enums/variants';
 import Label from '../Label';
 import { colors } from '../../index';
@@ -26,17 +26,19 @@ const generateCityList = (amount: number): OptionProps[] => {
 };
 const cities = generateCityList(1000);
 
-export const Basic: Story = args => {
+type BasicProps = DropdownProps & { clearable: boolean };
+
+export const Basic: Story<BasicProps> = ({ clearable, onClear, onSelect }: BasicProps) => {
   const [values, setValues] = useState<(string | number)[] | undefined>();
+
   return (
     <Label labelText="City" htmlFor="cities-list">
       <Dropdown
-        {...args}
         name="cities-list"
-        onClear={args.clearable ? args.onClear : undefined}
+        onClear={clearable ? onClear : undefined}
         onSelect={(selected?: Array<string | number>) => {
           setValues(selected);
-          return args.onSelect(selected);
+          return onSelect(selected);
         }}
         options={cities}
         values={values}