+++ src/components/TextInput/__tests__/TextInput.test.tsx
@@ -46,15 +46,15 @@ describe('TextInput', () => {
       );
       await waitFor(() => getByTestId(testId));
       expect(ref.current instanceof HTMLDivElement).toBeTruthy();
-    });      
+    });
     it('characterCountRef.current should exist', async () => {
       const ref = React.createRef<HTMLDivElement>();
       const { getByTestId } = render(
         <TextInput
-        showCharacterCount
-        maxLength={20}
-        characterCountRef={ref}
-        containerProps={containerProps}
+          showCharacterCount
+          maxLength={20}
+          characterCountRef={ref}
+          containerProps={containerProps}
         />,
       );
       await waitFor(() => getByTestId(testId));